msm: kgsl: Remove POPP
POPP constantly attempts to lower the GPU's frequency behind the governor's
back in order to save power; however, the GPU governor in use
(msm-adreno-tz) is very good at determining the GPU's load and selecting
an appropriate frequency to run the GPU at.

POPP was created long ago (perhaps when msm-adreno-tz didn't exist or
didn't work so well), so it is clearly deprecated. Remove it.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: ronaxdevil <pratabidya.007@gmail.com>
Change-Id: Ic712cc07b1715370ea93bfc087d4e9bc90c66df1